+++ src/util/helpers.rs
@@ -1,6 +1,5 @@
 use chrono::{DateTime, Utc};
 use colored::Colorize;
-use regex::Regex;
 use std::f64;
 use std::fs::{self, File};
 use std::io::{self, BufRead, Read};
@@ -284,13 +283,13 @@ pub fn get_fractional_number(flt_val: f64) -> i32 {
 * @param {String} text
 * @returns {String} description
 */
-pub fn get_description_from_text(text: &str) -> String {
-    let re = Regex::new(r"^(.*?)@.*").unwrap();
-    match re.captures(text) {
-        Some(mat) => mat[1].trim().parse().unwrap(),
-        None => text.to_string(),
-    }
-}
+// pub fn get_description_from_text(text: &str) -> String {
+//     let re = Regex::new(r"^(.*?)@.*").unwrap();
+//     match re.captures(text) {
+//         Some(mat) => mat[1].trim().parse().unwrap(),
+//         None => text.to_string(),
+//     }
+// }
 
 // ******************* UNIT TESTS *************
 
@@ -411,17 +410,17 @@ mod tests {
         assert_eq!(result, exp);
     }
 
-    #[test]
-    fn test_extract_with_with_tag() {
-        let text_01 = "This is a string with @tag";
-        let result = get_description_from_text(text_01);
-        assert_eq!(result, "This is a string with");
-    }
-
-    #[test]
-    fn test_extract_word_with_no_tag() {
-        let text_01 = "This is a string with no tag";
-        let result = get_description_from_text(text_01);
-        assert_eq!(result, "This is a string with no tag");
-    }
+    // #[test]
+    // fn test_extract_with_with_tag() {
+    //     let text_01 = "This is a string with @tag";
+    //     let result = get_description_from_text(text_01);
+    //     assert_eq!(result, "This is a string with");
+    // }
+    //
+    // #[test]
+    // fn test_extract_word_with_no_tag() {
+    //     let text_01 = "This is a string with no tag";
+    //     let result = get_description_from_text(text_01);
+    //     assert_eq!(result, "This is a string with no tag");
+    // }
 }